+++ src/scripts/events/saveEventListener.js
@@ -1,8 +1,18 @@
 // Matt Reeds - This is the event listener that gets the values of the added event and renders to dom
-import domOperations from "./eventsDomHandler.js"
-import apiActions from "./eventsDataHandler.js"
+import domOperations from "./events/eventsDomHandler.js"
+import apiActions from "./events/eventsDataHandler.js"
 
 const entryContainer = document.querySelector("#events-container");
+
+const clearForm = () => {
+    const eventName = document.getElementById("eventName");
+    const eventDate = document.getElementById("eventDate");
+    const eventLocation = document.getElementById("eventLocation");
+
+        eventName.value = "";
+        eventDate.value = "";
+        eventLocation.value= "";
+}
 
 const saveEventButton = {
     saveEventFunction: () => {
@@ -17,10 +27,11 @@ const saveEventButton = {
             const eventEntryObject = domOperations.eventEntryFactory(eventName, eventDate, eventLocation);
 
             apiActions.saveEventEntry(eventEntryObject)
-                    .then(() => {
-                        apiActions.getEvents()
-                            .then(domOperations.renderEventEntries)
-                    })
+                .then(() => {
+                    apiActions.getEvents()
+                        .then(domOperations.renderEventEntries)
+                        .then(clearForm)
+                })
         })
     }
 }
