+++ src/scripts/main.js
@@ -5,6 +5,7 @@ import apiActions from "./events/eventsDataHandler.js";
 import domOperations from "./events/eventsDomHandler.js";
 import newEventButton from "./events/eventFormListener.js";
 import saveEventButton from "./events/saveEventListener.js"
+import taskEvents from "./tasks/tasksEventListeners.js"
 
 /*
     Import all the tools into main.js that are needed to display
@@ -27,5 +28,4 @@ newEventButton.eventButtonFunction()
 saveEventButton.saveEventFunction()
 
 //tasks section tools - Cooper Nichols
-import taskEvents from "./tasks/tasksEventListeners.js"
 taskEvents.renderAllTasks()
